lib/config: rename to config-loader
It could be a bit confusing to potentially have lib/config and ./config
so let's rename the one we can.
