+++ lib/config-loader.js
