+++ lib/db.js
@@ -1,5 +1,5 @@
 var mysql = require('mysql'),
-    config = require('./config'),
+    config = require('./config-loader'),
     Promise = require('bluebird');
 
 const pool = mysql.createPool({
