+++ lib/git-manager.js
@@ -1,5 +1,5 @@
 var GithubApi = require('@octokit/rest'),
-    config = require('./config'),
+    config = require('./config-loader'),
     Promise = require('bluebird'),
     _ = require('underscore'),
     github = new GithubApi({
