+++ lib/pull-manager.js
@@ -1,6 +1,6 @@
 var debug      = require('./debug')('pulldasher:pull-manager'),
     pullQueue  = require('./pull-queue'),
-    config     = require('./config'),
+    config     = require('./config-loader'),
     _          = require('underscore'),
     dbManager  = require('./db-manager');
